groot/rtree: streamline use of rtree.ReadVarsFromStruct in scanners
Fixes go-hep/hep#641.
