+++ groot/rtree/scanner_test.go
@@ -1259,79 +1259,6 @@ func TestNewScanVars(t *testing.T) {
 	}
 }
 
-func TestInvalidScannerTypes(t *testing.T) {
-	t.Parallel()
-
-	fname := "../testdata/x-flat-tree.root"
-	f, err := riofs.Open(fname)
-	if err != nil {
-		t.Fatalf("could not open ROOT file %q: %v", fname, err)
-	}
-	defer f.Close()
-
-	obj, err := f.Get("tree")
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	tree := obj.(Tree)
-
-	for _, tc := range []struct {
-		name string
-		scan func() error
-		want error
-	}{
-		{
-			name: "scanner",
-			scan: func() error {
-				type Event struct {
-					N           int32  `groot:"N"`
-					notExported string `groot:"StdString"`
-				}
-				v := Event{
-					N:           42,
-					notExported: "N/A",
-				}
-				_, err := NewScanner(tree, &v)
-				return err
-			},
-			want: fmt.Errorf(`rtree: field[1] "notExported" from rtree.Event is not exported`),
-		},
-		{
-			name: "tree-scanner",
-			scan: func() error {
-				type Event struct {
-					N           int32  `groot:"N"`
-					notExported string `groot:"StdString"`
-				}
-				v := Event{
-					N:           42,
-					notExported: "N/A",
-				}
-				_, err := NewTreeScanner(tree, &v)
-				return err
-			},
-			want: fmt.Errorf(`rtree: field[1] "notExported" from rtree.Event is not exported`),
-		},
-	} {
-		t.Run(tc.name, func(t *testing.T) {
-			err := tc.scan()
-			switch {
-			case err == nil && tc.want == nil:
-				// ok
-			case err == nil && tc.want != nil:
-				t.Fatalf("invalid error: got no error, want=%v", tc.want)
-			case err != nil && tc.want == nil:
-				t.Fatalf("invalid error: got=%v, want=%v", err, tc.want)
-			default:
-				if got, want := err.Error(), tc.want.Error(); got != want {
-					t.Fatalf("invalid error:\ngot = %v\nwant= %v\n", got, want)
-				}
-			}
-		})
-	}
-}
-
 func TestG4LikeTree(t *testing.T) {
 	t.Parallel()
 	fname := rtests.XrdRemote("testdata/g4-like.root")
